fet(#2): Adjusting Solution structure
