--- source/Infraestrutura/ControleEstoque.Infraestrutura/Modulos/EntidadeBase.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace ControleEstoque.Infraestrutura.Modulos
-{
-    public abstract class EntidadeBase : EntidadeBaseComTipoId<long>
-    {
-    }
-}
