--- source/Infraestrutura/ControleEstoque.Infraestrutura/Modulos/IEntidadeBaseComTipoId.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace ControleEstoque.Infraestrutura.Modulos
-{
-    public interface IEntidadeBaseComTipoId<TId>
-    {
-        TId Id { get; }
-    }
-}
