--- source/Infraestrutura/ControleEstoque.Infraestrutura/Modulos/ObjetoValidavel.cs
@@ -1,26 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
-using System.Linq;
-using System.Net.Http.Headers;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace ControleEstoque.Infraestrutura.Modulos
-{
-    public abstract class ObjetoValidavel
-    {
-        public virtual bool EhValido()
-        {
-            return Validate().Count == 0;
-        }
-
-        public virtual IList<ValidationResult> Validate()
-        {
-            IList<ValidationResult> validationResults = new List<ValidationResult>();
-            Validator.TryValidateObject(this, new ValidationContext(this, null, null), validationResults, true);
-            return validationResults;
-        }
-        
-    }
-}
