--- source/Infraestrutura/ControleEstoque.Infraestrutura/Tarefas/Class1.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace ControleEstoque.Infraestrutura.Tarefas
-{
-    internal class Class1
-    {
-    }
-}
