+++ source/Infrastructure/StockControl.Infrastructure.Search/Class1.cs
@@ -1,0 +1,5 @@
+﻿namespace StockControl.Infrastructure.Search;
+public class Class1
+{
+
+}
