+++ source/Infrastructure/StockControl.Infrastructure.Search/StockControl.Infrastructure.Search.csproj
