+++ source/Infrastructure/StockControl.Infrastructure/Models/EntityBase.cs
@@ -1,0 +1,7 @@
+namespace StockControl.Infrastructure.Models
+{
+    public abstract class EntityBase : EntityBaseWithTypedId<long>
+    {
+        
+    }
+}
