+++ source/Infrastructure/StockControl.Infrastructure/Models/EntityBaseWithTypedId.cs
@@ -1,0 +1,8 @@
+namespace StockControl.Infrastructure.Models
+{
+    public abstract class EntityBaseWithTypedId<TId> : ValidatableObject, IEntityWithTypedId<TId>
+    {
+        public virtual TId Id {get; protected set;}
+        
+    }
+}
