+++ source/Infrastructure/StockControl.Infrastructure/Models/IEntityWithTypedId.cs
@@ -1,0 +1,7 @@
+namespace StockControl.Infrastructure.Models
+{
+    public interface IEntityWithTypedId<TId>
+    {
+        TId Id {get;}
+    }
+}
