+++ source/Infrastructure/StockControl.Infrastructure/Models/ValidatableObject.cs
@@ -1,0 +1,22 @@
+using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
+
+
+namespace StockControl.Infrastructure.Models
+{
+    public abstract class ValidatableObject
+    {
+        public virtual bool IsValid()
+        {
+            return Validate().Count == 0;
+        }
+
+        public virtual IList<ValidationResult> Validate()
+        {
+            IList<ValidationResult> validationResults = new List<ValidationResult>();
+            Validator.TryValidateObject(this, new ValidationContext(this, null, null), validationResults, true);
+            return validationResults;
+        }
+        
+    }
+}
