+++ source/Infrastructure/StockControl.Infrastructure/StockControl.Infrastructure.csproj
