+++ source/Modules/StockControl.Modules.Core/Class1.cs
@@ -1,0 +1,5 @@
+﻿namespace StockControl.Modules.Core;
+public class Class1
+{
+
+}
