+++ source/Modules/StockControl.Modules.Search/Class1.cs
@@ -1,0 +1,5 @@
+﻿namespace StockControl.Modules.Search;
+public class Class1
+{
+
+}
