--- source/Modulos/ControleEstoque.Modulo.Core/Extensoes/Class1.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace ControleEstoque.Modulo.Core.Extensoes
-{
-    internal class Class1
-    {
-    }
-}
