--- source/Web APIs/ControleEstoque.Servicos.Core.API/ControleEstoque.Servicos.Core.API.csproj.user
@@ -1,6 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="Current" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <PropertyGroup>
-    <ActiveDebugProfile>Docker</ActiveDebugProfile>
-  </PropertyGroup>
-</Project>
