--- source/Web APIs/ControleEstoque.Servicos.Core.API/WeatherForecast.cs
@@ -1,13 +1,0 @@
-namespace ControleEstoque.Servicos.Core.API
-{
-    public class WeatherForecast
-    {
-        public DateTime Date { get; set; }
-
-        public int TemperatureC { get; set; }
-
-        public int TemperatureF => 32 + (int)(TemperatureC / 0.5556);
-
-        public string? Summary { get; set; }
-    }
-}
