+++ source/Web APIs/StockControl.Services.Core.API/Program.cs
