+++ source/Web APIs/StockControl.Services.Core.API/appsettings.json
